requirements.txt
pandas
numpy
requests
python-dotenv
tavily-python
